+++ Classes/Service/CacheCleaner.php
@@ -1,14 +1,15 @@
 <?php
+
 declare(strict_types=1);
 
 namespace SourceBroker\Translatr\Service;
 
 use SourceBroker\Translatr\Utility\FileUtility;
-use SourceBroker\Translatr\Utility\MiscUtility;
 use TYPO3\CMS\Core\Cache\CacheManager;
 use TYPO3\CMS\Core\Cache\Exception\NoSuchCacheException;
-use TYPO3\CMS\Core\Http\HtmlResponse;
+use TYPO3\CMS\Core\Service\OpcodeCacheService;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Core\Utility\StringUtility;
 
 /**
  * Class CacheCleaner
@@ -31,21 +32,26 @@ class CacheCleaner extends BaseService
         $this->cacheManager = $this->objectManager->get(CacheManager::class);
     }
 
-    /**
-     * @return null|HtmlResponse
-     */
-    public function flushCache(): ?HtmlResponse
+    public function flushCache(): void
     {
-        $tempPath = FileUtility::getTempFolderPath();
-        $tempPathRenamed = $tempPath . time();
-        rename($tempPath, $tempPathRenamed);
-        GeneralUtility::rmdir($tempPathRenamed, true);
+        $directory = FileUtility::getTempFolderPath();
+        if (is_link($directory)) {
+            // Avoid attempting to rename the symlink see #87367
+            $directory = realpath($directory);
+        }
+        if (is_dir($directory)) {
+            $temporaryDirectory = rtrim($directory, '/') . '.' . StringUtility::getUniqueId('remove');
+            if (rename($directory, $temporaryDirectory)) {
+                GeneralUtility::makeInstance(OpcodeCacheService::class)->clearAllActive($directory);
+                GeneralUtility::mkdir($directory);
+                clearstatcache();
+                GeneralUtility::rmdir($temporaryDirectory, true);
+            }
+        }
         try {
             $cacheFrontend = $this->cacheManager->getCache('l10n');
             $cacheFrontend->flush();
         } catch (NoSuchCacheException $e) {
         }
-
-        return MiscUtility::isTypo39up() ? new HtmlResponse('') : null;
     }
 }