+++ ext_emconf.php
@@ -11,10 +11,10 @@ $EM_CONF[$_EXTKEY] = [
     'uploadfolder' => '1',
     'createDirs' => '',
     'clearCacheOnLoad' => 0,
-    'version' => '1.0.0',
+    'version' => '3.0.0',
     'constraints' => [
         'depends' => [
-            'typo3' => '8.7.0-9.5.999',
+            'typo3' => '9.5.999-10.4.999',
             'vhs' => '5.0.0-6.99.999',
         ],
         'conflicts' => [